remove use of tf.convert_image_dtype from model. we do this to support the fact we don't have this op available on NCS. changes expectation (everywhere) though that model input is float32 (-1.0 -> 1.0) instead of uint8 (0 -> 255) so needs changes in data pipeline as well as predict scripts

+++ generate_graph_pbtxt.py
@@ -14,7 +14,7 @@ parser.add_argument('--no-use-batch-norm', action='store_true')
 opts = parser.parse_args()
 
 # feed data through an explicit placeholder for frozen version
-imgs = tf.placeholder(dtype=tf.uint8, shape=(1, 1024, 768, 3), name='input_imgs')
+imgs = tf.placeholder(dtype=tf.float32, shape=(1, 1024, 768, 3), name='input_imgs')
 
 # restore model
 with tf.variable_scope("train_test_model") as scope:  # clumsy :/
